Combined keys, offsets, and making the shit into one class
