+++ test/test_helper.rb
@@ -3,7 +3,6 @@ SimpleCov.start
 require 'minitest/autorun'
 require 'minitest/pride'
 require 'mocha/minitest'
-require './lib/key'
-require './lib/offset'
+require './lib/key_offset'
 require './lib/shift'
 require './lib/enigma'
